requirements.txt
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==1.11.2
django-recaptcha==2.0.6
requests==2.25.1

[dev]
django-extensions>=1.5.9
django-debug-toolbar>=1.4

[tests]
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
mock==4.0.3
coverage==5.5
flake8==2.4.1
sphinx
six
